Fix transitive contentFiles support that's missing in nuget
So, turns out that nuget is completely broken for transitive assets that
aren't specifically `buildTransitive` assets. Why that scenario was fixed
just for that ultra-narrow use case of build assets is anyone's guess.

So, we leverage that one-and-only working transitive scenario to detect
the situation where transitive assets are missing (the Compile files we
provide as content) and include the missing files ourselves.

This requires changing the way we allow opt-out of the service attribute
type, though, since now there is no package reference to set the Exclude
on. For that, we provide a new property IncludeServiceAttribute that we
use to remove the automatically added contentFiles (in non-transitive)
or skip adding ourselves (in transitive).

Fixes #8