Replace ServiceAttribute inclusion via targets with source generator
The brokenness on the transitivity of contentFiles is tricky to overcome, and
proved more complicated to also maintain and test in various scenarios than
simply reading the same project property we have now, and skip adding the
static contents of the file via the source generator itself instead.

Adding a way to exclude the attribute is superfluous, so we just remove
that option. Users can always have their own attribute in any namespace.
Hopefully one day the [Service] attribute does come built-in the DI asm.

Fixes #12