Properly "exclude" service attribute when specified via props
The source generator we use to statically add the attribute before the
incremental generator runs, cannot check the analyzer options to skip
adding it. So it instead emits conditional code that checks for a
compiler defined constant which is set by the targets instead.

NOTE: we condition the constant to C# since that's the only language
we support for now, and VB constants need to be declared slightly
different too. Also would need tweaking if ever adding F# support.