Report warning when using the package without invoking AddServices
Conditions for reporting the warning are:

* Detect if users have access to IServiceCollection in the current project: it may be that
they just added the package to get the [Service] attribute for a library. Nothing to do
in that case.
* Detect if users actually invoke IServiceCollection members in the current project, more
or less flexibly (i.e. don't verify much where specifically, since the DI library can be used
in multiple contexts). If they do, then it's an indication that there should be a call to the
AddServices extension method too somewhere. As a hint, report the location on the
first invocation to a service collection method (or extension method).
* Report as a warning, since we can determine this wrong, depending on whether users
are initializing services some other way or indirectly. This allows them to ignore the warning
using the regular configuration mechanisms for analyzers and diagnostics.

Closes #30