Make development dependency
This package is intended for use just for either
getting the [Service] attribute, or invoking the
generated AddServices method.

In the former case, the class can be made
public (or even provided by another project),
so making this dependency transitive makes
no sense in that case.

In the second, the generated method must
also be the one in the referencing project,
so transitiveness for that case is also
unnecessary.

So make this project a pure development
dependency.

Closes #43