Add support for convention-based registrations
A new overload of `AddServices` allows passing a type and/or a regex to evaluate against types in the current compilation (at compile-time) and emit registrations for all of them, just as if they had been annotated with a corresponding `[Service]` attribute.

Key-based registrations not supported for this mechanism since it's not clear how to express a dynamic key based on either the type or regex that can be evaluated by the source generator.

Instead of adding a separte assembly-level attribute, extending the existing API with this pseudo-reflection approach is far more discoverable.

Fixes #116